[Upg] Updated song player to v14
