+++ .gitignore
@@ -74,3 +74,5 @@ typings/
 
 # FuseBox cache
 .fusebox/
+
+config.json
