+++ package.json
@@ -13,6 +13,7 @@
     "@discordjs/voice": "^0.6.0",
     "discord.js": "^14.9.0",
     "fs": "^0.0.1-security",
+    "libsodium-wrappers": "^0.7.11",
     "opusscript": "^0.0.8",
     "youtube-dl-exec": "^1.2.12",
     "yt-search": "^2.10.1",
