+++ src/Data/config.json
@@ -1,3 +1,6 @@
 {
-    "prefix": "!"
+    "prefix": "!",
+    "token": "ODg5NjUwMzM4MzQzMzEzNDY4.GTGR4f.WDbPfnS5ZxrIvpEhLiDmUWcLWTUwc4uoDmdrME",
+    "clientId": "889650338343313468",
+    "guildId": "125399154422841344"
 }
