+++ src/events/ready.js
@@ -1,0 +1,9 @@
+const { Events } = require('discord.js');
+
+module.exports = {
+    name: Events.ClientReady,
+    once: true,
+    execute(client){
+        console.log(`${client.user.tag} is ready to jam`);
+    }
+}
