--- src/scraper/scraperMain.js
@@ -1,8 +1,0 @@
-const browserObject = require('./browser');
-const scraperController = require('./pageController');
-
-let browserInstance = browserObject.startBrowser();
-
-const interval = setInterval(scraperController(browserInstance), 30000);
-
-module.exports ={interval};
